requirements.txt
agents==1.4.0
fastapi
uvicorn
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
deprecation==2.1.0
distro==1.9.0
fastapi==0.119.1
graphviz==0.21
griffe==1.14.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.17.0
mdurl==0.1.2
multidict==6.7.0
numpy==2.3.4
openai==1.109.1
openai-agents==0.3.3
packaging==25.0
pandas==2.3.3
postgrest==2.22.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
realtime==2.22.0
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
tqdm==4.67.1
typer==0.19.2
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
yarl==1.22.0